requirements.txt
pytesseract
pdf2image
pillow
tqdm
easyocr
pypdf
google-genai
customtkinter